Simple display rework (#10)
* Fixes bot remaining in voice channel

Bot remains in voice channel if final user switches from bot voice
channel to another voice channel

* Reworks display to provide different display options

`/sounds display` will display the following buttons

- `Display All`
- `Display Pinned`
- `Display Most Played`
- `Display Recently Added`

* Interacting with the buttons with display the appropriate buttons
below. The displays are capped at 50 buttons but can be altered with
environment vars.

* Adds Pin/Unpin capabilities for audio tracks to display favorites

* Enforces paginator limits (not the same as page limits)

* fixes incorrect version num (0.1.1 -> 0.2.0)

* reverses most_recently_added display buttons so newest is at bottom of screen

* adds reverse results for audio table paginator

The reverse results is important for situations in which order by is
used with pagination, and the results should be listed or displayed in
such a manor

* fixes incorrect reverse results when using limit and order by in audio table paginator

* adds description to clarify most recently added sound buttons layout

---------

Co-authored-by: krywicki <joseph.krywicki@gmail.com>